fix: move plugin-management file as it is not a mixin anymore
