+++ lib/main.js
@@ -5,10 +5,10 @@ import MinimapElement from './minimap-element'
 import Minimap from './minimap'
 import MinimapPluginGeneratorElement from './minimap-plugin-generator-element'
 import config from './config.json'
-import { deactivateAllPlugins } from './mixins/plugin-management'
+import { deactivateAllPlugins } from './plugin-management'
 
 exports.config = config
-export * from './mixins/plugin-management'
+export * from './plugin-management'
 export { default as Minimap } from './minimap'
 
 /**
