+++ lib/plugin-management.js
@@ -1,7 +1,7 @@
 'use strict'
 
 import { CompositeDisposable } from 'atom'
-import { emitter, getConfigSchema } from '../main'
+import { emitter, getConfigSchema } from './main'
 
 /**
  * Provides methods to manage minimap plugins.
